backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
requests>=2.31
